Split SpamEstimator into Bike/String/User submodules
Reorganize the flat SpamEstimator module into a namespace:
- SpamEstimator::String — the reusable text scorer (string_spaminess,
  looks_malicious?, within_bounds, suspiciousness helpers)
- SpamEstimator::Bike — estimate_bike and its bike/email/stolen-record helpers
- SpamEstimator::User — the former SeoSpaminess user-profile SEO-spam score

SpamEstimator::User now weights description and title far above name and
username (the latter scored at 0.2), since those are the SEO-spam payload.
